+++ src/OrchardCore.Modules/Orchard.Demo/Pages/_ViewImports.cshtml
@@ -1,4 +1,4 @@
-@inherits Orchard.DisplayManagement.Razor.Page
+@inherits Orchard.DisplayManagement.RazorPages.Page
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @addTagHelper *, Orchard.DisplayManagement
 @addTagHelper *, Orchard.ResourceManagement
