+++ src/OrchardCore/Orchard.Mvc.Core/RazorPages/ModularPageRouteModelConvention.cs
@@ -5,13 +5,14 @@ namespace Orchard.Mvc.RazorPages
     public class ModularPageRouteModelConvention : IPageRouteModelConvention
     {
         private const string PageFolder = "/Pages";
+        private const string PageSegment = PageFolder + "/";
 
         public void Apply(PageRouteModel model)
         {
             foreach (var selector in model.Selectors)
             {
                 selector.AttributeRouteModel.Template = selector.AttributeRouteModel
-                    .Template.Replace(PageFolder + "/", "/");
+                    .Template.Replace(PageSegment, "/");
             }
         }
     }
